Renamed send to sent, maximum value smallint in database is now short instead of int in code, fixed formatting so tabs are whitespaces, fixed and simplified multiplier calculation, added (value = <...>) to prevent classes renaming so they are unique for Sprint.
