+++ osgp-adapter-ws-distributionautomation/src/main/java/org/osgpfoundation/osgp/adapter/ws/da/application/config/ResendNotificationConfig.java
@@ -25,62 +25,62 @@ import com.alliander.osgp.shared.application.config.AbstractSchedulingConfig;
 @PropertySource(value = "file:${osgp/AdapterWsDistributionAutomation/config}", ignoreResourceNotFound = true)
 public class ResendNotificationConfig extends AbstractSchedulingConfig {
 
-	private static final String KEY_RESEND_NOTIFICATION_CRON_EXPRESSION = "distributionautomation.scheduling.job.resend.notification.cron.expression";
-	private static final String KEY_RESEND_NOTIFICATION_THREAD_COUNT = "distributionautomation.scheduling.job.resend.notification.thread.count";
+    private static final String KEY_RESEND_NOTIFICATION_CRON_EXPRESSION = "distributionautomation.scheduling.job.resend.notification.cron.expression";
+    private static final String KEY_RESEND_NOTIFICATION_THREAD_COUNT = "distributionautomation.scheduling.job.resend.notification.thread.count";
 
-	@Value("${db.driver}")
-	private String databaseDriver;
+    @Value("${db.driver}")
+    private String databaseDriver;
 
-	@Value("${db.password}")
-	private String databasePassword;
+    @Value("${db.password}")
+    private String databasePassword;
 
-	@Value("${db.protocol}")
-	private String databaseProtocol;
+    @Value("${db.protocol}")
+    private String databaseProtocol;
 
-	@Value("${db.host}")
-	private String databaseHost;
+    @Value("${db.host}")
+    private String databaseHost;
 
-	@Value("${db.port}")
-	private String databasePort;
+    @Value("${db.port}")
+    private String databasePort;
 
-	@Value("${db.name}")
-	private String databaseName;
+    @Value("${db.name}")
+    private String databaseName;
 
-	@Value("${db.username}")
-	private String databaseUsername;
+    @Value("${db.username}")
+    private String databaseUsername;
 
-	@Value("${distributionautomation.scheduling.job.resend.notification.cron.expression}")
-	private String resendNotificationCronExpression;
+    @Value("${distributionautomation.scheduling.job.resend.notification.cron.expression}")
+    private String resendNotificationCronExpression;
 
-	@Bean
-	public String resendNotificationCronExpression() {
-		return this.resendNotificationCronExpression;
-	}
+    @Value("${distributionautomation.scheduling.job.resend.notification.maximum}")
+    private short resendNotificationMaximum;
 
-	@Value("${distributionautomation.scheduling.job.resend.notification.maximum}")
-	private int resendNotificationMaximum;
+    @Value("${distributionautomation.scheduling.job.resend.notification.multiplier}")
+    private int resendNotificationMultiplier;
 
-	@Bean
-	public int resendNotificationMaximum() {
-		return this.resendNotificationMaximum;
-	}
+    @Bean
+    public String resendNotificationCronExpression() {
+        return this.resendNotificationCronExpression;
+    }
 
-	@Value("${distributionautomation.scheduling.job.resend.notification.multiplier}")
-	private int resendNotificationMultiplier;
+    @Bean
+    public short resendNotificationMaximum() {
+        return this.resendNotificationMaximum;
+    }
 
-	@Bean
-	public int resendNotificationMultiplier() {
-		return this.resendNotificationMultiplier;
-	}
+    @Bean
+    public int resendNotificationMultiplier() {
+        return this.resendNotificationMultiplier;
+    }
 
-	@Bean(destroyMethod = "shutdown")
-	public Scheduler resendNotificationScheduler() throws SchedulerException {
-		return this.constructScheduler(ResendNotificationJob.class, KEY_RESEND_NOTIFICATION_THREAD_COUNT,
-				KEY_RESEND_NOTIFICATION_CRON_EXPRESSION, this.getDatabaseUrl(), this.databaseUsername,
-				this.databasePassword, this.databaseDriver);
-	}
+    @Bean(destroyMethod = "shutdown")
+    public Scheduler resendNotificationScheduler() throws SchedulerException {
+        return this.constructScheduler(ResendNotificationJob.class, KEY_RESEND_NOTIFICATION_THREAD_COUNT,
+                KEY_RESEND_NOTIFICATION_CRON_EXPRESSION, this.getDatabaseUrl(), this.databaseUsername,
+                this.databasePassword, this.databaseDriver);
+    }
 
-	private String getDatabaseUrl() {
-		return this.databaseProtocol + this.databaseHost + ":" + this.databasePort + "/" + this.databaseName;
-	}
+    private String getDatabaseUrl() {
+        return this.databaseProtocol + this.databaseHost + ":" + this.databasePort + "/" + this.databaseName;
+    }
 }
