+++ osgp-adapter-ws-distributionautomation/src/main/java/org/osgpfoundation/osgp/adapter/ws/da/application/services/NotificationServiceWs.java
@@ -21,59 +21,59 @@ import com.alliander.osgp.shared.exceptionhandling.WebServiceSecurityException;
 @Validated
 public class NotificationServiceWs implements NotificationServiceDistributionAutomation {
 
-	private static final Logger LOGGER = LoggerFactory.getLogger(NotificationServiceWs.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(NotificationServiceWs.class);
 
-	private final SendNotificationServiceClient sendNotificationServiceClient;
+    private final SendNotificationServiceClient sendNotificationServiceClient;
 
-	private final String notificationUrl;
+    private final String notificationUrl;
 
-	private final String notificationUsername;
+    private final String notificationUsername;
 
-	private final String notificationOrganisation;
+    private final String notificationOrganisation;
 
-	public NotificationServiceWs(final SendNotificationServiceClient client, final String notificationUrl,
-			final String notificationUsername, final String notificationOrganisation) {
-		this.sendNotificationServiceClient = client;
-		this.notificationUrl = notificationUrl;
-		this.notificationUsername = notificationUsername;
-		this.notificationOrganisation = notificationOrganisation;
-	}
+    public NotificationServiceWs(final SendNotificationServiceClient client, final String notificationUrl,
+            final String notificationUsername, final String notificationOrganisation) {
+        this.sendNotificationServiceClient = client;
+        this.notificationUrl = notificationUrl;
+        this.notificationUsername = notificationUsername;
+        this.notificationOrganisation = notificationOrganisation;
+    }
 
-	/*
-	 * (non-Javadoc)
-	 *
-	 * @see com.alliander.osgp.adapter.ws.smartmetering.application.services.
-	 * INotificationService#sendNotification(java.lang.String, java.lang.String,
-	 * java.lang.String, java.lang.String, java.lang.String,
-	 * com.alliander.osgp.adapter
-	 * .ws.schema.smartmetering.notification.NotificationType)
-	 */
-	@Override
-	public void sendNotification(final String organisationIdentification, final String deviceIdentification,
-			final String result, final String correlationUid, final String message,
-			final NotificationType notificationType) {
+    /*
+     * (non-Javadoc)
+     *
+     * @see com.alliander.osgp.adapter.ws.smartmetering.application.services.
+     * INotificationService#sendNotification(java.lang.String, java.lang.String,
+     * java.lang.String, java.lang.String, java.lang.String,
+     * com.alliander.osgp.adapter
+     * .ws.schema.smartmetering.notification.NotificationType)
+     */
+    @Override
+    public void sendNotification(final String organisationIdentification, final String deviceIdentification,
+            final String result, final String correlationUid, final String message,
+            final NotificationType notificationType) {
 
-		LOGGER.info("sendNotification called with organisation: {}, correlationUid: {}, type: {}, to organisation: {}",
-				this.notificationOrganisation, correlationUid, notificationType, organisationIdentification);
+        LOGGER.info("sendNotification called with organisation: {}, correlationUid: {}, type: {}, to organisation: {}",
+                this.notificationOrganisation, correlationUid, notificationType, organisationIdentification);
 
-		final Notification notification = new Notification();
-		// message is null, unless an error occurred
-		notification.setMessage(message);
-		notification.setResult(result);
-		notification.setDeviceIdentification(deviceIdentification);
-		notification.setCorrelationUid(correlationUid);
-		notification.setNotificationType(notificationType);
+        final Notification notification = new Notification();
+        // message is null, unless an error occurred
+        notification.setMessage(message);
+        notification.setResult(result);
+        notification.setDeviceIdentification(deviceIdentification);
+        notification.setCorrelationUid(correlationUid);
+        notification.setNotificationType(notificationType);
 
-		try {
-			/*
-			 * Get a template for the organisation representing the OSGP platform, on behalf
-			 * of which the notification is sent to the organisation identified by the
-			 * organisationIdentification.
-			 */
-			this.sendNotificationServiceClient.sendNotification(this.notificationOrganisation, notification,
-					this.notificationUrl, this.notificationUsername);
-		} catch (final WebServiceSecurityException e) {
-			LOGGER.error(e.getMessage(), e);
-		}
-	}
+        try {
+            /*
+             * Get a template for the organisation representing the OSGP platform, on behalf
+             * of which the notification is sent to the organisation identified by the
+             * organisationIdentification.
+             */
+            this.sendNotificationServiceClient.sendNotification(this.notificationOrganisation, notification,
+                    this.notificationUrl, this.notificationUsername);
+        } catch (final WebServiceSecurityException e) {
+            LOGGER.error(e.getMessage(), e);
+        }
+    }
 }
