+++ osgp-adapter-ws-distributionautomation/src/main/resources/db/migration/V20180115125100777__Adds_number_of_notifications.sql
@@ -7,9 +7,9 @@ IF NOT EXISTS (
     FROM   information_schema.columns
     WHERE  table_schema = current_schema
     AND    table_name  = 'response_data'
-    AND	   column_name = 'number_of_notifications_send') THEN
+    AND	   column_name = 'number_of_notifications_sent') THEN
 
-    ALTER TABLE response_data ADD COLUMN number_of_notifications_send smallint NOT NULL DEFAULT(0);
+    ALTER TABLE response_data ADD COLUMN number_of_notifications_sent smallint NOT NULL DEFAULT(0);
 
 END IF;
