+++ osgp-adapter-ws-shared/src/main/java/com/alliander/osgp/adapter/ws/domain/repositories/ResponseDataRepository.java
@@ -18,15 +18,15 @@ import com.alliander.osgp.adapter.ws.domain.entities.ResponseData;
 @Repository
 public interface ResponseDataRepository extends JpaRepository<ResponseData, Long> {
 
-	List<ResponseData> findByOrganisationIdentification(String organisationIdentification);
+    List<ResponseData> findByOrganisationIdentification(String organisationIdentification);
 
-	List<ResponseData> findByMessageType(String messageType);
+    List<ResponseData> findByMessageType(String messageType);
 
-	List<ResponseData> findByDeviceIdentification(String deviceIdentification);
+    List<ResponseData> findByDeviceIdentification(String deviceIdentification);
 
-	List<ResponseData> findByNumberOfNotificationsSendLessThan(Integer numberOfNotificationsSend);
+    List<ResponseData> findByNumberOfNotificationsSentLessThan(Short numberOfNotificationsSent);
 
-	ResponseData findByCorrelationUid(String correlationUid);
+    ResponseData findByCorrelationUid(String correlationUid);
 
-	void removeByCreationTimeBefore(Date date);
+    void removeByCreationTimeBefore(Date date);
 }
