+++ osgp-adapter-ws-smartmetering/src/main/java/com/alliander/osgp/adapter/ws/smartmetering/infra/jms/messageprocessor/DomainResponseMessageProcessor.java
@@ -36,139 +36,139 @@ import com.alliander.osgp.shared.infra.jms.ResponseMessageResultType;
  */
 public abstract class DomainResponseMessageProcessor implements MessageProcessor {
 
-	/**
-	 * Logger for this class.
-	 */
-	private static final Logger LOGGER = LoggerFactory.getLogger(DomainResponseMessageProcessor.class);
+    /**
+     * Logger for this class.
+     */
+    private static final Logger LOGGER = LoggerFactory.getLogger(DomainResponseMessageProcessor.class);
 
-	/**
-	 * The hash map of message processor instances.
-	 */
-	@Autowired
-	protected DomainResponseMessageProcessorMap domainResponseMessageProcessorMap;
+    /**
+     * The hash map of message processor instances.
+     */
+    @Autowired
+    protected DomainResponseMessageProcessorMap domainResponseMessageProcessorMap;
 
-	@Autowired
-	private NotificationService notificationService;
+    @Autowired
+    private NotificationService notificationService;
 
-	@Autowired
-	private ResponseDataService responseDataService;
+    @Autowired
+    private ResponseDataService responseDataService;
 
-	/**
-	 * The message type that a message processor implementation can handle.
-	 */
-	protected DeviceFunction deviceFunction;
+    /**
+     * The message type that a message processor implementation can handle.
+     */
+    protected DeviceFunction deviceFunction;
 
-	/**
-	 * Construct a message processor instance by passing in the message type.
-	 *
-	 * @param deviceFunction
-	 *            The message type a message processor can handle.
-	 */
-	protected DomainResponseMessageProcessor(final DeviceFunction deviceFunction) {
-		this.deviceFunction = deviceFunction;
-	}
+    /**
+     * Construct a message processor instance by passing in the message type.
+     *
+     * @param deviceFunction
+     *            The message type a message processor can handle.
+     */
+    protected DomainResponseMessageProcessor(final DeviceFunction deviceFunction) {
+        this.deviceFunction = deviceFunction;
+    }
 
-	/**
-	 * Initialization function executed after dependency injection has finished. The
-	 * MessageProcessor Singleton is added to the HashMap of MessageProcessors. The
-	 * key for the HashMap is the integer value of the enumeration member.
-	 */
-	@PostConstruct
-	public void init() {
-		this.domainResponseMessageProcessorMap.addMessageProcessor(this.deviceFunction.ordinal(),
-				this.deviceFunction.name(), this);
-	}
+    /**
+     * Initialization function executed after dependency injection has finished. The
+     * MessageProcessor Singleton is added to the HashMap of MessageProcessors. The
+     * key for the HashMap is the integer value of the enumeration member.
+     */
+    @PostConstruct
+    public void init() {
+        this.domainResponseMessageProcessorMap.addMessageProcessor(this.deviceFunction.ordinal(),
+                this.deviceFunction.name(), this);
+    }
 
-	@Override
-	public void processMessage(final ObjectMessage message) throws JMSException {
-		LOGGER.debug("Processing smart metering response message");
+    @Override
+    public void processMessage(final ObjectMessage message) throws JMSException {
+        LOGGER.debug("Processing smart metering response message");
 
-		String correlationUid = null;
-		String messageType = null;
-		String organisationIdentification = null;
-		String deviceIdentification = null;
+        String correlationUid = null;
+        String messageType = null;
+        String organisationIdentification = null;
+        String deviceIdentification = null;
 
-		String notificationMessage = null;
-		NotificationType notificationType = null;
-		ResponseMessageResultType resultType = null;
-		String resultDescription = null;
-		Serializable dataObject = null;
+        String notificationMessage = null;
+        NotificationType notificationType = null;
+        ResponseMessageResultType resultType = null;
+        String resultDescription = null;
+        Serializable dataObject = null;
 
-		try {
-			correlationUid = message.getJMSCorrelationID();
-			messageType = message.getJMSType();
-			organisationIdentification = message.getStringProperty(Constants.ORGANISATION_IDENTIFICATION);
-			deviceIdentification = message.getStringProperty(Constants.DEVICE_IDENTIFICATION);
-			resultType = ResponseMessageResultType.valueOf(message.getStringProperty(Constants.RESULT));
-			resultDescription = message.getStringProperty(Constants.DESCRIPTION);
+        try {
+            correlationUid = message.getJMSCorrelationID();
+            messageType = message.getJMSType();
+            organisationIdentification = message.getStringProperty(Constants.ORGANISATION_IDENTIFICATION);
+            deviceIdentification = message.getStringProperty(Constants.DEVICE_IDENTIFICATION);
+            resultType = ResponseMessageResultType.valueOf(message.getStringProperty(Constants.RESULT));
+            resultDescription = message.getStringProperty(Constants.DESCRIPTION);
 
-			notificationMessage = message.getStringProperty(Constants.DESCRIPTION);
-			notificationType = NotificationType.valueOf(messageType);
+            notificationMessage = message.getStringProperty(Constants.DESCRIPTION);
+            notificationType = NotificationType.valueOf(messageType);
 
-			dataObject = message.getObject();
-		} catch (final JMSException e) {
-			LOGGER.error("UNRECOVERABLE ERROR, unable to read ObjectMessage instance, giving up.", e);
-			LOGGER.debug("correlationUid: {}", correlationUid);
-			LOGGER.debug("messageType: {}", messageType);
-			LOGGER.debug("organisationIdentification: {}", organisationIdentification);
-			LOGGER.debug("deviceIdentification: {}", deviceIdentification);
-			return;
-		}
+            dataObject = message.getObject();
+        } catch (final JMSException e) {
+            LOGGER.error("UNRECOVERABLE ERROR, unable to read ObjectMessage instance, giving up.", e);
+            LOGGER.debug("correlationUid: {}", correlationUid);
+            LOGGER.debug("messageType: {}", messageType);
+            LOGGER.debug("organisationIdentification: {}", organisationIdentification);
+            LOGGER.debug("deviceIdentification: {}", deviceIdentification);
+            return;
+        }
 
-		try {
-			LOGGER.info("Calling application service function to handle response: {} with correlationUid: {}",
-					messageType, correlationUid);
+        try {
+            LOGGER.info("Calling application service function to handle response: {} with correlationUid: {}",
+                    messageType, correlationUid);
 
-			this.handleMessage(organisationIdentification, messageType, deviceIdentification, correlationUid,
-					resultType, resultDescription, dataObject);
+            this.handleMessage(organisationIdentification, messageType, deviceIdentification, correlationUid,
+                    resultType, resultDescription, dataObject);
 
-			// Send notification indicating data is available.
-			this.notificationService.sendNotification(organisationIdentification, deviceIdentification,
-					resultType.name(), correlationUid, notificationMessage, notificationType);
+            // Send notification indicating data is available.
+            this.notificationService.sendNotification(organisationIdentification, deviceIdentification,
+                    resultType.name(), correlationUid, notificationMessage, notificationType);
 
-		} catch (final Exception e) {
-			this.handleError(e, correlationUid, organisationIdentification, deviceIdentification, notificationType);
-		}
-	}
+        } catch (final Exception e) {
+            this.handleError(e, correlationUid, organisationIdentification, deviceIdentification, notificationType);
+        }
+    }
 
-	protected void handleMessage(final String organisationIdentification, final String messageType,
-			final String deviceIdentification, final String correlationUid, final ResponseMessageResultType resultType,
-			final String resultDescription, final Serializable dataObject) {
+    protected void handleMessage(final String organisationIdentification, final String messageType,
+            final String deviceIdentification, final String correlationUid, final ResponseMessageResultType resultType,
+            final String resultDescription, final Serializable dataObject) {
 
-		final int numberOfNotificationsSend = 0;
-		Serializable meterResponseObject;
-		if (dataObject == null) {
-			meterResponseObject = resultDescription;
-		} else {
-			meterResponseObject = dataObject;
-		}
+        final short numberOfNotificationsSent = 0;
+        Serializable meterResponseObject;
+        if (dataObject == null) {
+            meterResponseObject = resultDescription;
+        } else {
+            meterResponseObject = dataObject;
+        }
 
-		final ResponseData responseData = new ResponseData(organisationIdentification, messageType,
-				deviceIdentification, correlationUid, resultType, meterResponseObject, numberOfNotificationsSend);
-		this.responseDataService.enqueue(responseData);
-	}
+        final ResponseData responseData = new ResponseData(organisationIdentification, messageType,
+                deviceIdentification, correlationUid, resultType, meterResponseObject, numberOfNotificationsSent);
+        this.responseDataService.enqueue(responseData);
+    }
 
-	/**
-	 * In case of an error, this function can be used to send a response containing
-	 * the exception to the web-service-adapter.
-	 *
-	 * @param e
-	 *            The exception.
-	 * @param correlationUid
-	 *            The correlation UID.
-	 * @param organisationIdentification
-	 *            The organisation identification.
-	 * @param deviceIdentification
-	 *            The device identification.
-	 * @param notificationType
-	 *            The message type.
-	 */
-	protected void handleError(final Exception e, final String correlationUid, final String organisationIdentification,
-			final String deviceIdentification, final NotificationType notificationType) {
+    /**
+     * In case of an error, this function can be used to send a response containing
+     * the exception to the web-service-adapter.
+     *
+     * @param e
+     *            The exception.
+     * @param correlationUid
+     *            The correlation UID.
+     * @param organisationIdentification
+     *            The organisation identification.
+     * @param deviceIdentification
+     *            The device identification.
+     * @param notificationType
+     *            The message type.
+     */
+    protected void handleError(final Exception e, final String correlationUid, final String organisationIdentification,
+            final String deviceIdentification, final NotificationType notificationType) {
 
-		LOGGER.info("handeling error: {} for notification type: {} with correlationUid: {}", e.getMessage(),
-				notificationType, correlationUid, e);
-		this.notificationService.sendNotification(organisationIdentification, deviceIdentification, "NOT_OK",
-				correlationUid, e.getMessage(), notificationType);
-	}
+        LOGGER.info("handeling error: {} for notification type: {} with correlationUid: {}", e.getMessage(),
+                notificationType, correlationUid, e);
+        this.notificationService.sendNotification(organisationIdentification, deviceIdentification, "NOT_OK",
+                correlationUid, e.getMessage(), notificationType);
+    }
 }
